# Conflicts:
#	MiniAssignment4.cpp
#	Scene.cpp

+++ Sphere.h
@@ -7,10 +7,10 @@ class Sphere : public SceneObject {
 public:
 	using SceneObject::intersect;
 	using SceneObject::setMaterial;
-	Sphere(Eigen::Vector3f centre, int radius);
+	Sphere(Eigen::Vector3f centre, float radius);
 	~Sphere();
 	Eigen::Vector3f centre;
-	int radius;
+	float radius;
 	Intersection intersect(Ray ray);
 };
